refactor: separate binaries into each package and facade package
This commit separates each binary into modules that are easier to understand and test.

Closes https://github.com/Evaneos/vitaminjs/issues/396

refactor: var changed to const

chore: remove es6 modules after pulling master

chore: remove useless es5 binary and separate test command
