+++ packages/vitaminjs-build/bin/listenExitSignal.js
@@ -1,0 +1,5 @@
+module.exports = (callback) => {
+    ['SIGINT', 'SIGTERM', 'SIGQUIT'].forEach((signal) => {
+        process.on(signal, () => callback(signal));
+    });
+};
