+++ packages/vitaminjs-build/bin/setVitaminPath.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const path = require('path');
+
+process.env.VITAMIN_PATH = path.resolve(__dirname, '..');
