+++ packages/vitaminjs-build/bin/vitamin-clean.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+const clean = require('./clean');
+
+require('./setVitaminPath');
+clean();
