+++ packages/vitaminjs-build/package.json
@@ -17,7 +17,10 @@
     "url": "https://github.com/Evaneos/vitaminjs.git"
   },
   "bin": {
-    "vitamin": "bin/vitamin.js"
+    "vitamin-build": "bin/vitamin-build.js",
+    "vitamin-clean": "bin/vitamin-clean.js",
+    "vitamin-start": "bin/vitamin-start.js",
+    "vitamin-test": "bin/vitamin-test.js"
   },
   "dependencies": {
     "babel-core": "^6.24.1",
