+++ packages/vitaminjs-runtime/package.json
@@ -16,9 +16,13 @@
     "type": "git",
     "url": "https://github.com/Evaneos/vitaminjs.git"
   },
+  "bin": {
+    "vitamin-serve": "bin/vitamin-serve.js"
+  },
   "main": "src/shared/index.js",
   "dependencies": {
     "chalk": "^1.1.3",
+    "commander": "^2.12.2",
     "express": "^4.15.2",
     "history": "^3.2.1",
     "http-graceful-shutdown": "^1.0.6",
