+++ packages/vitaminjs/bin/vitamin-version.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const { version } = require('../package.json');
+
+process.stdout.write(`v${version}\n`);
