--- packages/vitaminjs/bin/vitamin.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-module.exports = require('vitaminjs-build/bin/vitamin.js');
